Backend/requirements.txt
alembic==1.11.2
APScheduler==3.10.1
attrs==19.3.0
backports.zoneinfo==0.2.1
boto3==1.28.20
botocore==1.31.20
certifi==2019.11.28
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
colorhash==1.2.1
configparser==6.0.0
cycler==0.10.0
databricks-cli==0.17.7
docker==6.1.3
docx2pdf==0.1.7
docx2txt==0.8
entrypoints==0.4
Flask==2.2.5
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
from-root==1.3.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==3.2.0
kiwisolver==1.1.0
kneed==0.5.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib
mlflow==1.30.1
more-itertools==8.1.0
mysql-connector-python==8.0.33
#mysqlclient
nltk==3.6.2
numpy==1.21.6
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
pathspec==0.11.2
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
#protobuf
psutil==5.9.5
PyJWT==2.8.0
pyparsing==3.1.1
#pypiwin32==223
pyrsistent==0.15.7
python-dateutil==2.8.2
pytz
#pywin32
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2021.4.4
regexp==0.1
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.60.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
waitress==2.1.2
websocket-client==1.6.1
Werkzeug==2.2.3
wincertstore==0.2
xgboost==0.90
yamllint==1.28.0
zipp==3.15.0
streamlit
pymongo
